+++ src/com/jpexs/decompiler/flash/gui/MainFrameRibbon.java
@@ -54,29 +54,17 @@ public final class MainFrameRibbon extends AppRibbonFrame {
 
     public MainFrameRibbon() {
         super();
-
-        FlashPlayerPanel flashPanel = null;
-        FlashPlayerPanel flashPanel2 = null;
-
-        if (Configuration.useAdobeFlashPlayerForPreviews.get()) {
-            try {
-                flashPanel = new FlashPlayerPanel(this);
-                flashPanel2 = new FlashPlayerPanel(this);
-            } catch (FlashUnsupportedException fue) {
-                //ignored
-            }
-        }
+        
 
         Container cnt = getContentPane();
         cnt.setLayout(new BorderLayout());
         JRibbon ribbon = getRibbon();
         cnt.add(ribbon, BorderLayout.NORTH);
 
-        boolean externalFlashPlayerUnavailable = flashPanel == null;
-        mainMenu = new MainFrameRibbonMenu(this, ribbon, externalFlashPlayerUnavailable);
+        mainMenu = new MainFrameRibbonMenu(this, ribbon);
         mainMenu.createMenuBar();
 
-        panel = new MainPanel(this, mainMenu, flashPanel, flashPanel2);
+        panel = new MainPanel(this, mainMenu);
         panel.setBackground(Color.yellow);
         cnt.add(panel, BorderLayout.CENTER);